move aggregation_by_rel utility to utils.py
