+++ src/kg_topology_toolbox/topology_toolbox.py
@@ -5,15 +5,16 @@
 Topology toolbox main functionalities
 """
 
-from collections.abc import Iterable
-from functools import cache
 import warnings
+from functools import cache
+
 import numpy as np
 import pandas as pd
 from pandas.api.types import is_integer_dtype
 from scipy.sparse import coo_array
 
 from kg_topology_toolbox.utils import (
+    aggregate_by_relation,
     composition_count,
     jaccard_similarity,
     node_degrees_and_rels,
@@ -281,7 +282,9 @@ class KGTopologyToolbox:
             ).astype(str)
         return df_res
 
-    def edge_degree_cardinality_summary(self) -> pd.DataFrame:
+    def edge_degree_cardinality_summary(
+        self, aggregate_by_r: bool = False
+    ) -> pd.DataFrame:
         """
         For each edge in the KG, compute the number of edges with the same head
         (head-degree, or out-degree), the same tail (tail-degree, or in-degree)
@@ -293,6 +296,10 @@ class KGTopologyToolbox:
 
         The output dataframe maintains the same indexing and ordering of triples
         as the original Knowledge Graph dataframe.
+
+        :param aggregate_by_r:
+            If True, return metrics aggregated by relation type
+            (the output DataFrame will be indexed over relation IDs).
 
         :return:
             The results dataframe. Contains the following columns
@@ -344,16 +351,17 @@ class KGTopologyToolbox:
         df_res["triple_cardinality_same_rel"] = edge_cardinality[
             "triple_cardinality_same_rel"
         ]
-        return df_res
+        return aggregate_by_relation(df_res) if aggregate_by_r else df_res
 
     def edge_pattern_summary(
         self,
         return_metapath_list: bool = False,
         composition_chunk_size: int = 2**8,
         composition_workers: int = 32,
+        aggregate_by_r: bool = False,
     ) -> pd.DataFrame:
         """
-        This function analyses the structural properties of each edge in the graph:
+        Analyse structural properties of each edge in the KG:
         symmetry, presence of inverse/inference(=parallel) edges and
         triangles supported on the edge.
 
@@ -368,10 +376,13 @@ class KGTopologyToolbox:
             to compute the triangle count.
         :param composition_workers:
             Number of workers to compute the triangle count.
+        :param aggregate_by_r:
+            If True, return metrics aggregated by relation type
+            (the output DataFrame will be indexed over relation IDs).
 
         :return:
             The results dataframe. Contains the following columns
-            (in addition to `h`, `r`, `t` in ``df``):
+            (in addition to `h`, `r`, `t`):
 
             - **is_loop** (bool): True if the triple is a loop (``h == t``).
             - **is_symmetric** (bool): True if the triple (t, r, h) is also contained
@@ -506,7 +517,7 @@ class KGTopologyToolbox:
         )
         df_res["has_undirected_composition"] = df_res["n_undirected_triangles"] > 0
 
-        return df_res[
+        df_res = df_res[
             [
                 "h",
                 "r",
@@ -527,84 +538,7 @@ class KGTopologyToolbox:
             + (["metapath_list"] if return_metapath_list else [])
         ]
 
-    def aggregate_by_relation(self, edge_topology_df: pd.DataFrame) -> pd.DataFrame:
-        """
-        Aggregate topology metrics of all triples of the same relation type.
-        To be applied to the output dataframe of either
-        :meth:`KGTopologyToolbox.edge_degree_cardinality_summary` or
-        :meth:`KGTopologyToolbox.edge_pattern_summary`.
-
-        The returned dataframe is indexed over relation type IDs, with columns
-        giving the aggregated statistics of triples of the correspondig relation.
-        The name of the columns is of the form ``column_name_in_input_df + suffix``.
-        The aggregation is perfomed by returning:
-
-        - for numerical metrics: mean, standard deviation and quartiles
-          (``suffix`` = "_mean", "_std", "_quartile1", "_quartile2", "_quartile3");
-        - for boolean metrics: the fraction of triples of the relation type
-          with metric = True (``suffix`` = "_frac");
-        - for string metrics: for each possible label, the fraction of triples
-          of the relation type with that metric value (``suffix`` = "_{label}_frac")
-        - for list metrics: the unique metric values across triples of the relation
-          type (``suffix`` = "_unique").
-
-        :param edge_topology_df: pd.DataFrame of edge topology metrics.
-            Must contain at least three columns `h`, `r`, `t`.
-
-        :return: The results dataframe. In addition to the columns with the aggregated
-            metrics by relation type, it also contains columns:
-
-            - **num_triples** (int): Number of triples for each relation type.
-            - **frac_triples** (float): Fraction of overall triples represented by each
-              relation type.
-            - **unique_h** (int): Number of unique head entities used by triples of each
-              relation type.
-            - **unique_t** (int): Number of unique tail entities used by triples of each
-              relation type.
-        """
-        df_by_r = edge_topology_df.groupby("r")
-        df_res = df_by_r.agg(num_triples=("r", "count"))
-        df_res["frac_triples"] = df_res["num_triples"] / edge_topology_df.shape[0]
-        col: str
-        for col, col_dtype in edge_topology_df.drop(columns=["r"]).dtypes.items():  # type: ignore
-            if col in ["h", "t"]:
-                df_res[f"unique_{col}"] = df_by_r[col].nunique()
-            elif col_dtype == object:
-                if isinstance(edge_topology_df[col].iloc[0], str):
-                    for label in np.unique(edge_topology_df[col]):
-                        df_res[f"{col}_{label}_frac"] = (
-                            edge_topology_df[edge_topology_df[col] == label]
-                            .groupby("r")[col]
-                            .count()
-                            / df_res["num_triples"]
-                        ).fillna(0)
-                elif isinstance(edge_topology_df[col].iloc[0], Iterable):
-                    df_res[f"{col}_unique"] = (
-                        df_by_r[col]
-                        .agg(np.unique)
-                        .apply(
-                            lambda x: (
-                                np.unique(
-                                    np.concatenate(
-                                        [lst for lst in x if len(lst) > 0] or [[]]
-                                    )
-                                ).tolist()
-                            )
-                        )
-                    )
-                else:
-                    print(f"Skipping column {col}: no known aggregation mode")
-                    continue
-            elif col_dtype == int or col_dtype == float:
-                df_res[f"{col}_mean"] = df_by_r[col].mean()
-                df_res[f"{col}_std"] = df_by_r[col].std()
-                for q in range(1, 4):
-                    df_res[f"{col}_quartile{q}"] = df_by_r[col].agg(
-                        lambda x: np.quantile(x, 0.25 * q)
-                    )
-            elif col_dtype == bool:
-                df_res[f"{col}_frac"] = df_by_r[col].mean()
-        return df_res
+        return aggregate_by_relation(df_res) if aggregate_by_r else df_res
 
     def jaccard_similarity_relation_sets(self) -> pd.DataFrame:
         """
